Fix base16 task exec order and set task
Prior to this change the base16 task were executed before the vim tasks.
This breaks the automation because vim reads the configuration and tries
to set the base16 colors before the base16 vim plugin is installed. Now
the vim configuration cannot find `~/.vimrc_background` and loads
without issues.

In addition prior to this change the `Set base16_monokai theme` ansible
task stops the ansible playbook execution because it spawns an
interactive shell. By adding the `exit` command it sets the theme and
exits from the shell.